requirements.txt
numpy==1.26.4
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
